requirements/testing.txt
# PyTest for running the tests.
pytest==2.9.1
pytest-django==2.9.1
pytest-cov==1.8.1